refactor(searchPage): split component
