+++ packages/mirror-tv-next/README.md
@@ -1,4 +1,4 @@
 # mirror-tv-next
 
 ## 專案文件
-請見此[Document](https://paper.dropbox.com/doc/Conventions--CRV1VF0ZmlU4yOPOlpCdnKJUAQ-Y5dSxqXJ16cTuQlOgg6LB)
+請見此[Document](https://paper.dropbox.com/doc/Mirror-tv-next-Tabris--CRXwBOTzPQCfpv4Qf_xx8CSQAQ-wcxZtqLMJAWbZ8RZ8F053)
