+++ packages/mirror-tv-next/app/search/[keyword]/page.tsx
@@ -1,11 +1,6 @@
 import NoSearchResult from '~/components/search/no-search-result'
-import type {
-  PopularSearchItem,
-  TVPost,
-  TVPostResponse,
-} from '~/types/api-data'
+import type { TVPost, TVPostResponse } from '~/types/api-data'
 import SearchResult from '~/components/search/search-result'
-import { fetchPopularPosts } from '~/components/errors/action'
 
 type Slug = {
   params: Params
@@ -27,7 +22,6 @@ const getSearchResult = async (keyword?: string): Promise<TVPostResponse> => {
 const page = async ({ params }: Slug) => {
   const keyword = decodeURI(params.keyword)
   let searchResultList: TVPost[] = []
-  let popularPostList: PopularSearchItem[] = []
   try {
     const searchResultResponse = await getSearchResult('')
     searchResultList = searchResultResponse.body.hits.hits || []
@@ -35,17 +29,7 @@ const page = async ({ params }: Slug) => {
     console.error('searchResultList_error', e)
   }
 
-  if (searchResultList.length === 0) {
-    try {
-      const { data } = await fetchPopularPosts()
-      popularPostList = data.report
-    } catch (e) {
-      console.error('popularPostList_error', e)
-    }
-    return (
-      <NoSearchResult keyword={keyword} popularPostList={popularPostList} />
-    )
-  }
+  if (searchResultList.length === 0) return <NoSearchResult keyword={keyword} />
 
   const searchResultProps = {
     keyword,
