+++ packages/mirror-tv-next/components/search/_styles/no-search-result-list.module.scss
@@ -1,0 +1,29 @@
+@import '/styles/theme/breakpoints';
+
+.popularResultList {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  align-items: center;
+  gap: 28px;
+  @include media-breakpoint-up(md) {
+    display: grid;
+    grid-template-columns: repeat(3, 1fr);
+    gap: 20px;
+  }
+
+  @include media-breakpoint-up(lg) {
+    grid-template-columns: repeat(4, 1fr);
+    gap: 24px;
+  }
+
+  a {
+    display: block;
+    width: 100%;
+    height: 100%;
+    li {
+      width: 100%;
+      height: 100%;
+    }
+  }
+}
